Move nested constant parser classes to own files
